Refactored data flows
The flow of data through the application and UI components was not done in a way that was easy to reason about.  This was partially due to unpredictable event flows and multiple ways data gets bound.

This commit changes that to keep a consistent flow, where all changes to emitters or application state are all propagated through a single sequence.  AppOperations are enqueued to signal changes, the application runs through each operation, applies them to the application state, and the application state is then propagated up through all UI components.