Add pooling of particles
Particle pools allow a large batch of particles to be allocated once and shared among multiple emitters.  When an emitter is created it reserves a slice of particles, and when that emitter is disposed those particles are freed up for the next emitter to claim. The intention is to relieve memory pressure with short duration emitters.