Standardized color values
This standardizes coloring of particles where RGB values are between 0 and 255 instead of 0.0 and 1.0.  This is more natural for most designers and artists.  However, alpha is still 0.0 to 1.0 as it represents opaque percentage.

This commit also removes the color change modifier as it wasn't intuitive to think about how the color multiplier changes every second.  Instead it was replaced with the `EndingColorMultiplier`, which allows linear interpolation between a starting color multiplier and a final color multiplier based on the particle's lifetime.

Closes #19