Fixed crash when going from texture to no texture emitter
When working with an emitter that has a texture and then loading a texture that did not have a texture file set, the editor would crash as the texture section editor was trying to load a texture from a `null` file.  This occurred because the texture section display editor was not removing previous texture information when going from texture to no texture.

This also updated the texture file editor to clean itself up to when going from texture to no-texture, so the UI doesn't endd up in a confused state.