Fix one shot triggers
One shot triggers was setting `_hasTriggered` even if the emitter wasn't set to create new particles.  This meant that one-shot triggered emitters that were not set to create particles immediately on the first `Update()` call would never have particles created.  The use of `_hasTriggered` to prevent more than one emission also made it impossible to know if the emitter was still emitting new particles.  Technically the answer was "no" but it will still return true.

So this fixes both of those issues by removing the `_hasTriggered` variable and just set the emitter to no longer be emitting after the update cycle is finished.  This also makes it so the emitter can be restarted by just setting `IsEmittingNewParticles` to `true`.