Don't copy `Parme.Frb.dll` to glue plugin directory
Glue automatically tries to load all dlls in the plugin folder.  However, `Parme.Frb.dll` causes glue to fail to load the plugin since that dll requires monogame to be referenced as well.

The plugin itself doesn't actually need that dll in order to run, so it has been removed from the copy list.