Added rotation to emitters
Emitters can now have rotational values set to them.  When a particle is created (after all initailizers have run) the new particles' rotation, velocity, and positional offsets will be rotated by the emitter's rotation value.  This allows effects to be consistent when angled differently, so for instance a wedge contrails on a bullet fired at different angles.

This can be opted out of by just not adjusting the rotation on emitters that this is not desired on.