Fixed code gen culture issue
When code generation occurs on a system whose culture is set to one that uses comma for decimal place separators, the code that's generated is not valid C# numbers.  This fixes the case by making sure that all generated code is being formatted as invariant culture instead of the current culture.