Fix ending colors not iterating properly
Since a particle's current color is a byte it was losing precision, and thus would only work properly when going from 255 to 0.  So the Current properties were changed to float to fix this.

The code generation for the ending color modifier was also setting the ending alpha incorrectly, as the modifier itself is a float between 0.0 to 1.0 but the property needs to be between 0 and 255.  Starting color initializer's code generation modified this as expected but ending did not, so this has been fixed.