Fixed issue where emitter logic types would not be found in glue.
Glue does not automatically remove ATIs added by a plugin when the plugin is shut down,. so when the plugin is loaded a second time the Parme plugin was adding a new batch of ATIs.  When this occurs all emitter logic types are added to the new ATIs, but glue references the old ATIs with empty emitter logic types,  and therefore displays no emitter logic types to the user.

This fixes it by making sure all ATIs are cleaned up when the plugin is shut down.