Add setting to control copying of textures.
Currently if the user selects a texture that is not in the same directory as the emlogic file then it will ask if they want to copy the texture to be in the same directory.  This can get annoying if the user is always having to select no, because they want a specific directory layout (e.g. texture in the parent directory of particles).

By changing this option the user will no longer be asked if they want to copy the file over, and each user can have their preference saved.

Closes #51